Reproducible example
An alternative to the `spatialdata.to_polygons()` label vectorization  function with the extra functionality to ensure that the geometries are valid. Additionally, there is a possibility to filter out individual geoms to constitute `shapely.MultiPolygon` based on the geom area. This implementation is ~2x faster compare to the SpatialData probably due to not chunking the label image. Also simplifies the process so that there is no need to create `SpatialData()` prior running vectorization.